ドメインモデル図更新 モデルも併せて修正
